Claude IA 2

+++ src/main/java/dev/rodolfo/sistema_advogados/repository/AdvogadoRepository.java
@@ -3,6 +3,6 @@ package dev.rodolfo.sistema_advogados.repository;
 import dev.rodolfo.sistema_advogados.entity.Advogado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdvogadoRepository extends JpaRepository<Advogado, Integer> {
+public interface AdvogadoRepository extends JpaRepository<Advogado, Long> {
     Advogado findByOAB(String OAB);
 }
